# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小胰宝智能助手平台",
    "description": "基于多AI平台的智能体系统，支持用户认证和权限管理",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/api-keys/": {
      "post": {
        "tags": [
          "API Key管理"
        ],
        "summary": "创建API Key",
        "description": "创建新的API Key（需要高级用户权限）",
        "operationId": "create_api_key_api_v1_api_keys__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyFullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "聊天"
        ],
        "summary": "Get Chat",
        "description": "聊天API（仅支持API Key认证）\n\n此端点专门为API Key用户设计，提供稳定的聊天服务。\n用户需要先登录获取JWT令牌，然后使用JWT创建API Key，最后使用API Key访问此接口。",
        "operationId": "get_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "指定智能体名称。也支持 'Agent'/'x-agent' 变体。示例: fastgpt"
          }
        ]
      }
    },
    "/api/v1/upload": {
      "post": {
        "tags": [
          "文件上传"
        ],
        "summary": "Upload File",
        "description": "文件上传API（仅支持API Key认证）\n\n此端点专门为API Key用户设计，提供稳定的文件上传服务。\n用户需要先登录获取JWT令牌，然后使用JWT创建API Key，最后使用API Key访问此接口。\n\n参数:\n    file: 上传的文件\n    agent: 指定的智能体名称（可选，也可通过header传递）\n    created_by: 创建者标识（可选）\n\n返回:\n    FileUploadResponse: 包含文件信息的响应",
        "operationId": "upload_file_api_v1_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "可选智能体名称（也支持表单字段 'agent'，以及 'Agent'/'x-agent' 变体）"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "API Key名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "API Key描述"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "过期时间，为空表示永不过期"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "权限列表",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreate",
        "description": "创建API Key请求模型"
      },
      "APIKeyFullResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix",
            "description": "API Key前缀（用于识别）"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "description": "创建者用户ID"
          },
          "full_key": {
            "type": "string",
            "title": "Full Key",
            "description": "完整的API Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "key_prefix",
          "created_at",
          "expires_at",
          "last_used_at",
          "permissions",
          "is_active",
          "created_by",
          "full_key"
        ],
        "title": "APIKeyFullResponse",
        "description": "API Key完整响应模型（包含完整密钥）"
      },
...
      "Body_upload_file_api_v1_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v1_upload_post"
      },
      "ChatRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "用户输入的对话内容，所有Agent必填"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User",
            "description": "用户标识"
          },
          "uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uid",
            "description": "用户ID"
          },
          "stream": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream",
            "description": "是否流式返回",
            "default": false
          },
          "app_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Id",
            "description": "应用ID"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id",
            "description": "会话ID"
          },
          "files": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files",
            "description": "上传的文件列表"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "会话标识，用于多轮对话",
            "default": ""
          },
          "detail": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "是否返回详细信息（模块状态等）",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ChatRequest",
        "description": "前端请求统一数据模型\n\n说明:\n    - query: 所有 Agent 必须要求前端传递，表示用户输入的对话内容（必填）\n    - 其他字段为可选，各 Agent 可按需提取，在 Agent 中需要对该必填字段进行校验"
      },
      "Choice": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/Message",
            "description": "消息对象"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason",
            "description": "结束原因"
          },
          "index": {
            "type": "integer",
            "title": "Index",
            "description": "选项索引"
          }
        },
        "type": "object",
        "required": [
          "message",
          "index"
        ],
        "title": "Choice"
      },
      "CompleteMessage": {
        "properties": {
          "obj": {
            "type": "string",
            "title": "Obj",
            "description": "消息对象类型，如 System, Human, AI"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "消息内容"
          }
        },
        "type": "object",
        "required": [
          "obj",
          "value"
        ],
        "title": "CompleteMessage",
        "description": "完整消息"
      },
      "FileUploadResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "响应代码，0表示成功"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "文件信息数据"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "响应消息",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "FileUploadResponse",
        "description": "文件上传响应模型"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "消息角色，如 assistant 或 user"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "消息内容"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "Message"
      },
...
      "QuoteItem": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "数据集ID"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "项目ID"
          },
          "q": {
            "type": "string",
            "title": "Q",
            "description": "问题"
          },
          "a": {
            "type": "string",
            "title": "A",
            "description": "答案"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "来源"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "id",
          "q",
          "a",
          "source"
        ],
        "title": "QuoteItem",
        "description": "引用项"
      },
      "ResponseDataModule": {
        "properties": {
          "moduleName": {
            "type": "string",
            "title": "Modulename",
            "description": "模块名称"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price",
            "description": "模块价格"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "使用的模型"
          },
          "tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens",
            "description": "tokens 数量"
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity",
            "description": "相似度"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit",
            "description": "限制数量"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question",
            "description": "问题"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer",
            "description": "答案"
          },
          "maxToken": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxtoken",
            "description": "最大 token 数"
          },
          "quoteList": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/QuoteItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quotelist",
            "description": "引用列表"
          },
          "completeMessages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CompleteMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completemessages",
            "description": "完整消息列表"
          },
          "runningTime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runningtime",
            "description": "运行时间"
          },
          "totalPoints": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalpoints",
            "description": "总点数"
          },
          "inputTokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputtokens",
            "description": "输入 tokens"
          },
          "outputTokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputtokens",
            "description": "输出 tokens"
          },
          "reasoningText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoningtext",
            "description": "推理文本"
          },
          "historyPreview": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historypreview",
            "description": "历史预览"
          },
          "contextTotalLen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexttotallen",
            "description": "上下文总长度"
          },
          "finishReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishreason",
            "description": "完成原因"
          },
          "nodeId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nodeid",
            "description": "节点ID"
          },
          "moduleType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moduletype",
            "description": "模块类型"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "模块ID"
          }
        },
        "type": "object",
        "required": [
          "moduleName"
        ],
        "title": "ResponseDataModule",
        "description": "响应数据模块 - FastGPT detail=true 时的模块信息"
      },
...
      "UnifiedChatResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "会话ID"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "使用的模型名称"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ],
            "description": "tokens 使用情况"
          },
          "choices": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Choice"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choices",
            "description": "响应选项列表"
          },
          "responseData": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ResponseDataModule"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsedata",
            "description": "详细的模块响应数据"
          },
          "newVariables": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newvariables",
            "description": "新变量"
          }
        },
        "type": "object",
        "title": "UnifiedChatResponse",
        "description": "统一聊天响应格式 - 支持两种模式"
      },
      "Usage": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "description": "提示词 tokens 数"
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "description": "生成 tokens 数"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "description": "总 tokens 数"
          }
        },
        "type": "object",
        "required": [
          "prompt_tokens",
          "completion_tokens",
          "total_tokens"
        ],
        "title": "Usage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}